chatbot/requirements.txt
<<<<<<< HEAD
discord.py>=2.3.2
python-dotenv>=0.19.0
requests>=2.31.0
cohere>=4.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
aiohttp>=3.8.0
openai>=0.28.0
pydantic>=1.10.0
python-multipart>=0.0.5
jinja2>=3.0.0
=======
discord.py>=2.3.2
python-dotenv>=0.19.0
requests>=2.31.0
cohere>=4.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
aiohttp>=3.8.0
openai>=0.28.0
pydantic>=1.10.0
python-multipart>=0.0.5
jinja2>=3.0.0

>>>>>>> 71be574313c6a352c72e5658a8bffc248098cdcc
